Making layout, fragment and viewModel for loginScreen including pin code(confirmation code) stage.

+++ app/src/main/java/app/cashadvisor/common/domain/model/ErrorEntity.kt
@@ -28,7 +28,7 @@ sealed class ErrorEntity(open val message: String) {
         data class WrongConfirmationCode(
             override val message: String,
             val remainingAttempts: Int,
-            val lockDuration: Int
+            val lockDuration: Long
         ) : LoginConfirmationWithCode(message)
 
         data class FailedToConfirmEmailOrLoginUser(override val message: String) :
@@ -43,7 +43,7 @@ sealed class ErrorEntity(open val message: String) {
         data class WrongConfirmationCode(
             override val message: String,
             val remainingAttempts: Int,
-            val lockDuration: Int
+            val lockDuration: Long
         ) : RegisterConfirmationWithCode(message)
 
         data class FailedToConfirmEmailOrRegisterUser(override val message: String) :
